@@ -1,8 +1,9 @@
 {
     "name": "innovcamp2024sveltesurge",
     "version": "0.0.0",
+    "private": true,
     "type": "module",
-    "packageManager": "pnpm@9.1.0",
+    "packageManager": "pnpm@9.12.2+sha512.22721b3a11f81661ae1ec68ce1a7b879425a1ca5b991c975b074ac220b187ce56c708fe5db69f4c962c989452eee76c82877f4ee80f474cebd61ee13461b6228",
     "scripts": {
         "dev": "vite",
         "build": "vite build",
@@ -10,45 +11,45 @@
         "sync": "svelte-kit sync",
         "fmt": "prettier --check .",
         "fmt:fix": "prettier --write .",
-        "lint": "concurrently pnpm:lint:*",
+        "lint": "pnpm --parallel /^lint:*/",
         "lint:html": "linthtml src/**/*.html",
         "lint:css": "stylelint src/**/*.css",
         "lint:js": "eslint **/*.{js,svelte,ts}",
         "lint:svelte": "svelte-check --tsconfig ./tsconfig.json"
     },
     "devDependencies": {
-        "@linthtml/linthtml": "^0.9.6",
+        "@linthtml/linthtml": "^0.10.1",
         "@linthtml/linthtml-config-recommended": "^0.1.0",
-        "@skeletonlabs/skeleton": "^2.10.0",
         "@skeletonlabs/tw-plugin": "^0.4.0",
-        "@sveltejs/adapter-static": "^3.0.1",
-        "@sveltejs/kit": "^2.5.7",
-        "@sveltejs/vite-plugin-svelte": "^3.1.0",
-        "@tailwindcss/typography": "^0.5.13",
-        "@typescript-eslint/eslint-plugin": "^7.8.0",
-        "@typescript-eslint/parser": "^7.8.0",
-        "autoprefixer": "^10.4.19",
-        "concurrently": "^8.2.2",
-        "eslint": "^8.57.0",
+        "@sveltejs/adapter-static": "^3.0.5",
+        "@sveltejs/kit": "^2.7.2",
+        "@sveltejs/vite-plugin-svelte": "^4.0.0",
+        "@tailwindcss/typography": "^0.5.15",
+        "autoprefixer": "^10.4.20",
+        "eslint": "^9.13.0",
         "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-import": "^2.29.1",
-        "eslint-plugin-svelte": "^2.39.0",
-        "htmlnano": "^2.1.1",
-        "prettier": "^3.2.5",
-        "prettier-plugin-svelte": "^3.2.3",
-        "prettier-plugin-tailwindcss": "^0.5.14",
-        "stylelint": "^16.5.0",
-        "stylelint-config-standard": "^36.0.0",
+        "eslint-plugin-import": "^2.31.0",
+        "eslint-plugin-svelte": "^2.46.0",
+        "globals": "^15.11.0",
+        "prettier": "^3.3.3",
+        "prettier-plugin-svelte": "^3.2.7",
+        "prettier-plugin-tailwindcss": "^0.6.8",
+        "stylelint": "^16.10.0",
+        "stylelint-config-standard": "^36.0.1",
         "stylelint-config-tailwindcss": "^0.0.7",
-        "svelte": "^4.2.15",
-        "svelte-check": "^3.7.1",
-        "tailwindcss": "^3.4.3",
-        "terser": "^5.31.0",
-        "typescript": "^5.4.5",
-        "vite": "^5.2.11",
+        "svelte-check": "^4.0.5",
+        "tailwindcss": "^3.4.14",
+        "terser": "^5.36.0",
+        "typescript": "^5.6.3",
+        "typescript-eslint": "^8.11.0",
+        "vite": "^5.4.10",
         "vite-plugin-tailwind-purgecss": "^0.3.3"
     },
     "dependencies": {
-        "@fontsource/bebas-neue": "^5.0.20"
+        "@fontsource/bebas-neue": "^5.1.0",
+        "@skeletonlabs/skeleton": "^2.10.3",
+        "htmlnano": "^2.1.1",
+        "svelte": "^5.1.2",
+        "svelte-eslint-parser": "^0.43.0"
     }
 }
